Refactor. Remove dead code. Move to appropriate package
