+++ server/src/main/scala/services/graphstages/DynamicWorkloadCalculator.scala
@@ -6,86 +6,15 @@ import drt.shared.FlightsApi.FlightsWithSplits
 import drt.shared._
 import drt.shared.api.Arrival
 import org.slf4j.{Logger, LoggerFactory}
-import services.SDate
 import services.graphstages.Crunch.{FlightSplitMinute, SplitMinutes}
 import services.workloadcalculator.PaxLoadCalculator.Load
-import uk.gov.homeoffice.drt.egates.PortEgateBanksUpdates
 import uk.gov.homeoffice.drt.ports.Queues._
 import uk.gov.homeoffice.drt.ports.Terminals.Terminal
-import uk.gov.homeoffice.drt.ports.{AirportConfig, ApiPaxTypeAndQueueCount, PaxTypeAndQueue, Queues}
+import uk.gov.homeoffice.drt.ports.{ApiPaxTypeAndQueueCount, PaxTypeAndQueue, Queues}
 import uk.gov.homeoffice.drt.redlist.RedListUpdates
 
-import scala.collection.immutable
-import scala.collection.immutable.{Map, NumericRange}
-import scala.concurrent.{ExecutionContext, Future}
-
-object QueueStatusProviders {
-
-  trait QueueStatusProvider {
-    def statusAt(terminal: Terminal, queue: Queue, time: SDateLike): Future[QueueStatus]
-  }
-
-  case object QueuesAlwaysOpen extends QueueStatusProvider {
-    override def statusAt(terminal: Terminal, queue: Queue, time: SDateLike): Future[QueueStatus] = Future.successful(Open)
-  }
-
-  case class DynamicQueueStatusProvider(airportConfig: AirportConfig, egatesProvider: () => Future[PortEgateBanksUpdates])
-                                       (implicit ec: ExecutionContext) {
-
-    def allStatusesForPeriod: NumericRange[MillisSinceEpoch] => Future[Map[Terminal, Map[Queue, Map[MillisSinceEpoch, QueueStatus]]]] =
-      period => {
-        egatesProvider().map { egatePortUpdates =>
-          airportConfig.queuesByTerminal.map { case (terminal, queues) =>
-            val byQueue = queues.map {
-              case EGate => (EGate, egateStatuses(period, egatePortUpdates, terminal))
-              case queue => (queue, deskStatuses(airportConfig.maxDesksByTerminalAndQueue24Hrs, period, terminal, queue))
-            }.toMap
-            (terminal, byQueue)
-          }
-        }
-      }
-  }
-
-  private def egateStatuses(period: NumericRange[MillisSinceEpoch],
-                            egatePortUpdates: PortEgateBanksUpdates,
-                            terminal: Terminal): Map[MillisSinceEpoch, QueueStatus] =
-    egatePortUpdates.updatesByTerminal
-      .get(terminal)
-      .map { updates =>
-        val statuses = updates.forPeriod(period).map(b => if (b.exists(!_.isClosed)) Open else Closed)
-        period.zip(statuses)
-      }
-      .getOrElse(period.map(m => (m, Closed)))
-      .toMap
-
-  private def deskStatuses(maxDesks: Map[Terminal, Map[Queue, IndexedSeq[Int]]],
-                           period: NumericRange[MillisSinceEpoch],
-                           terminal: Terminal,
-                           queue: Queue): Map[MillisSinceEpoch, QueueStatus] =
-    maxDesks
-      .get(terminal)
-      .flatMap(_.get(queue).map { maxByHour =>
-        val maxByHourLifted = maxByHour.lift
-        period.map { minute =>
-          val status = maxByHourLifted(SDate(minute).getHours()) match {
-            case None => Closed
-            case Some(0) => Closed
-            case Some(_) => Open
-          }
-          (minute, status)
-        }
-      })
-      .getOrElse(period.map(m => (m, Closed)))
-      .toMap
-
-  case class FlexibleEGatesForSimulation(eGateOpenHours: Seq[Int]) extends QueueStatusProvider {
-    override def statusAt(t: Terminal, queue: Queue, time: SDateLike): Future[QueueStatus] =
-      queue match {
-        case EGate if !eGateOpenHours.contains(time.getHours()) => Future.successful(Closed)
-        case _ => Future.successful(Open)
-      }
-  }
-}
+import scala.collection.immutable.Map
+import scala.concurrent.ExecutionContext
 
 
 trait WorkloadCalculatorLike {
