+++ server/src/main/scala/services/scenarios/Scenarios.scala
@@ -13,12 +13,12 @@ import drt.shared.SimulationParams
 import drt.shared.api.Arrival
 import manifests.queues.SplitsCalculator
 import passengersplits.parsing.VoyageManifestParser
+import queueus.DynamicQueueStatusProvider
 import services.crunch.desklimits.PortDeskLimits
 import services.crunch.deskrecs.DynamicRunnableDeskRecs.HistoricManifestsProvider
 import services.crunch.deskrecs.RunnableOptimisation.CrunchRequest
 import services.crunch.deskrecs.{DynamicRunnableDeskRecs, PortDesksAndWaitsProvider, RunnableOptimisation}
 import services.graphstages.FlightFilter
-import services.graphstages.QueueStatusProviders.DynamicQueueStatusProvider
 import services.{OptimiserWithFlexibleProcessors, SDate}
 import uk.gov.homeoffice.drt.egates.PortEgateBanksUpdates
 import uk.gov.homeoffice.drt.ports.AirportConfig
